scale-simp-webapp/requirements.txt
fastapi==0.118.1
uvicorn==0.37.0
pydantic==2.11.9
openai==2.1.0
scikit-learn==1.6.1
umap-learn==0.5.9.post2
hdbscan==0.8.40
bertopic==0.17.3
pandas==2.3.1
numpy==2.2.6
matplotlib==3.10.3
scipy==1.16.1